code arrangment
