+++ .vscode/launch.json
@@ -1,0 +1,13 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "name": "Scouting_App_Off_Season",
+            "request": "launch",
+            "type": "dart"
+        }
+    ]
+}
