+++ lib/main.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:scouting_app_2/Pages/Home/Home.dart';
-import 'package:scouting_app_2/Pages/Login/widgets/LoginScreen.dart';
 import 'package:scouting_app_2/services/firebase_auth_service.dart';
+
+import 'Pages/Home/Home.dart';
+import 'Pages/Login/widgets/LoginScreen.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
